+++ HISTORY.md
@@ -2,6 +2,6 @@
 
 This branch is fork of [IsmaelMartinez teams-for-linux](https://github.com/mahmoudbahaa/outlook-for-linux). but replacing teams with outlook see that repo for actual history
 
--- Updated Icon from Teams to Outlook
--- Updated Url to open outlook.office.com
--- Updated various md and config files to point to outlook instead of teams
+- Updated Icon from Teams to Outlook
+- Updated Url to open outlook.office.com
+- Updated various md and config files to point to outlook instead of teams